Move ArrayUtils detail code into separate header
